pyproject.toml
[tool.poetry]
name = "my-budget-pal-bot"
version = "0.1.0"
description = "Telegram Mini App for personal budget tracking."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "budget_bot", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.10.0"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
aiosqlite = "^0.21.0"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
sqlmodel = "^0.0.24"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.23.7"
ruff = "^0.12.9"
mypy = "^1.10.0"
bandit = "^1.7.9"
radon = "^6.0.1"
pre-commit = "^4.3.0"
pytest-cov = "^6.2.1"
httpx = "^0.28.1"
pip-audit = "^2.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.scripts]
start = "budget_bot.main:run_main"

# --- НАСТРОЙКИ ИНСТРУМЕНТОВ КАЧЕСТВА ---

[tool.ruff]
line-length = 88

[tool.ruff.lint]
# Включаем стандартные правила Pyflakes, Pycodestyle и Isort
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
plugins = ["sqlalchemy.ext.mypy.plugin"]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false
